Add cmake and 7zip packages to /buildbot with a script to unpack
The base chromium buildbot image lacks some dependencies that we use to
build DR and drmemory, in particular cmake and 7zip.  This will allow us
to automatically install these on all of our bots.

The tools won't auto-update if we check a new archive into svn.  If we
need to force an update, we can clobber the bot's build directory.


SVN-Revision: 1080